Improved argument handling and configuration file. Fixes #263
I've used the argument handling used in cvsanaly and apply it to bicho. Now
we have more options (debug mode, version, optional delay) and the
configuration file is similar to the one used by cvsanaly.

I've also divided the utils.py functions and the Config class. This part is
very similar in cvsanaly.