Fix #334. Launchpad shared bugs are now assigned to different trackers.
Due to bugs can be shared, a bug can have more than a related
tracker. This information is now stored in the issues table with
new trackers created every time a new one appears. This is useful
when user passes as an url a meta tracker which contains a list of
bugs shared among different projects.