[Bugzilla] Fix problem getting last changes of issues
Bugzilla returns the changes from a date, including those
changes made on that date. Due to this, we need to increase
the date of the last modification adding one second, to get
new changes or issues and to not retrieve those already stored.
This commit is critical to fix #22 because there woudn't be a way
to know whether we have retrieve all issues or not. Moreover,
it helps to fix issue #20.