[BackendManager] New structure for backends
The structure of backends has changed. Now, backends will be
implemented as Python packages and have to be coded on __init__.py
file.
BackendManager will look for packages in the given directory,
importing those backends from __init__.py files. This new structure
provides a more flexible and clean organization of the backends code.