[Bugzilla] Add Bugzilla issues summary parser
In Bugzilla, The full list of issues can be retrieved from a CSV file
that includes the identifier of each bug and the date when the last
change was made. This list will be used as reference for crawling the
full issue tracker site.

The information from each issue is unmarshalled in a new class named
BugzillaIssueSummary.