[Redmine] Add Redmine identities parser
Identities or 'users' in Redmine can be retrieved parsing a JSON
stream. This stream includes additional info such as the login name
of the user, the date when the user was created or the last time
the user accessed the tracker. The user id is an int and not the
login attrib.

The information from each user is unmarshalled to RedmineIdentity
class.