[Launchpad] Skip private issues when are duplicates
Fixes #127.

When an issue is private in Lanchpad and you do not have enough
access privileges, the server returns with a HTTP 404 error.
To fix the problem a "try/except" block has been added to catch
the error, skipping the access to the private issue. This was needed
because there was no other way to check when an issue is private
(without access privileges to that issue).