[Bugzilla] Fix bug detecting proper changes table
Fixes #134.

In Bugzilla, the changes table has 5 'th' tags (or columns)
which include the names of each field ('Who', 'When', etc).
To detect the changes table we searched for the first table
with 5 'th' tags. This search was recursive, so when a table
has child tables (including the changes tables), we were finding
it but we were selecting the parent table for parsing instead
of the changes table.

Including the parameter 'recursive=False' when looking for 'th'
tags made the work.